# src/ui/iPath.Blazor.Server/wwwroot/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "iPath.Blazor.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/taskassignments/my": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserTaskAssignmentsQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultListOfTaskAssignmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taskassignments/group/{groupId}": {
      "get": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "statusFilter",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/eTaskStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/case/{serviceRequestId}": {
      "get": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "serviceRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}": {
      "get": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/propose": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeTaskAssignmentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}/accept": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}/decline": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}/complete": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}/return": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/{id}/cancel": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/taskassignments/followup": {
      "post": {
        "tags": [
          "Task Assignments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowUpTaskCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodedConcept": {
        "type": "object",
        "properties": {
          "system": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "display": {
            "type": "string"
          }
        }
      },
...
      "CreateFollowUpTaskCommand": {
        "required": [
          "serviceRequestId"
        ],
        "type": "object",
        "properties": {
          "serviceRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "eTaskAssignmentMode": {
        "type": "integer"
      },
...
      "eTaskStatus": {
        "type": "integer"
      },
...
      "GetUserTaskAssignmentsQuery": {
        "type": "object",
        "properties": {
          "userId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "statusFilter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/eTaskStatus"
              }
            ]
          },
          "includeServiceRequest": {
            "type": "boolean"
          },
          "sorting": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "filter": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 10
          }
        }
      },
...
      "OwnerDto": {
        "required": [
          "id",
          "username",
          "email"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultListOfTaskAssignmentDto": {
        "required": [
          "totalItems",
          "items"
        ],
        "type": "object",
        "properties": {
          "totalItems": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskAssignmentDto"
            }
          }
        }
      },
...
      "PatientInfo": {
        "type": "object",
        "properties": {
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "gender": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProposeTaskAssignmentCommand": {
        "required": [
          "serviceRequestId",
          "assignedToUserId",
          "mode"
        ],
        "type": "object",
        "properties": {
          "serviceRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "assignedToUserId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/eTaskAssignmentMode"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "QuestionnaireResponseData": {
        "type": "object",
        "properties": {
          "questionnaireId": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resource": {
            "type": [
              "null",
              "string"
            ]
          },
          "generatedText": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RequestDescription": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "storage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StorageInfo"
              }
            ]
          },
          "subtitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "caseType": {
            "type": [
              "null",
              "string"
            ]
          },
          "accessionNo": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "minLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "text": {
            "type": [
              "null",
              "string"
            ]
          },
          "provisionalDiagnosis": {
            "type": [
              "null",
              "string"
            ]
          },
          "provisionalDiagnosisCode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CodedConcept"
              }
            ]
          },
          "patientInfo": {
            "$ref": "#/components/schemas/PatientInfo"
          },
          "questionnaire": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/QuestionnaireResponseData"
              }
            ]
          },
          "bodySite": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CodedConcept"
              }
            ]
          },
          "isClinicalInfoValid": {
            "type": "boolean"
          }
        }
      },
...
      "ServiceRequestListDto": {
        "required": [
          "owner"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nodeType": {
            "type": "string"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "isDraft": {
            "type": "boolean"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerDto"
          },
          "groupId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RequestDescription"
              }
            ]
          },
          "annotationCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastVisit": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastAnnotationDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "StorageInfo": {
        "type": "object",
        "properties": {
          "providerName": {
            "type": "string"
          },
          "storageId": {
            "type": "string"
          },
          "updatedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "TaskAssignmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "serviceRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "groupName": {
            "type": [
              "null",
              "string"
            ]
          },
          "assignedToUserId": {
            "type": "string",
            "format": "uuid"
          },
          "assignedToUsername": {
            "type": [
              "null",
              "string"
            ]
          },
          "assignedByUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "assignedByUsername": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "completedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "serviceRequest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServiceRequestListDto"
              }
            ]
          }
        }
      },
...
    }
  },
...
}